+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Changes
+- Removed dependency on Architectury API.
 ## Fixes
-- Fixed sea level not being set properly for chunk provider when entering chunk provider settings
-- Fixed crash when opening chunk provider settings.
+- Fixed Beta water colours not working on NeoForge
+- Fixed Mangrove Leaves not having Beta leaf colouring
+- Fixed Bush, Pink Petal and Wildflower blocks not having Beta grass colouring.
+- Fixed Tall Grass and Large Fern blocks not having consistent colouring between the two halves.